# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/createpod/{podName}/{image}": {
            "post": {
                "description": "Create a pod with the given image name in Kubernetes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a pod",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pod Name",
                        "name": "podName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image Name",
                        "name": "image",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pod created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid image name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create pod",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/exposepod/{podName}/{port}": {
            "post": {
                "description": "Expose a pod with the given pod name in Kubernetes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Expose a pod",
                "operationId": "expose-pod",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pod Name",
                        "name": "podName",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Port Number",
                        "name": "port",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pod exposed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid pod name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to expose pod",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}